# docs/swagger.yaml
basePath: /
definitions:
...
  handler.NewPassword:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    type: object
  handler.ProfileData:
    properties:
      date_of_birth:
        type: string
      name:
        type: string
      number:
        type: string
      password:
        type: string
    type: object
...
  models.AgeCategory:
    properties:
      id:
        type: integer
      max_age:
        type: integer
      min_age:
        type: integer
    type: object
  models.Cover:
    properties:
      filename:
        type: string
      id:
        type: integer
      project_id:
        type: integer
    type: object
  models.Episode:
    properties:
      episode_number:
        type: integer
      id:
        type: integer
      link:
        type: string
      season_id:
        type: integer
    type: object
  models.Genre:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Keyword:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.Screenshot:
    properties:
      filename:
        type: string
      id:
        type: integer
      project_id:
        type: integer
    type: object
  models.Season:
    properties:
      episodes:
        items:
          $ref: '#/definitions/models.Episode'
        type: array
      id:
        type: integer
      season_number:
        type: integer
      series_id:
        type: integer
    type: object
  models.Series:
    properties:
      age_categories:
        items:
          $ref: '#/definitions/models.AgeCategory'
        type: array
      cover:
        $ref: '#/definitions/models.Cover'
      description:
        type: string
      director:
        type: string
      duration:
        type: integer
      genres:
        items:
          $ref: '#/definitions/models.Genre'
        type: array
      id:
        type: integer
      keywords:
        items:
          $ref: '#/definitions/models.Keyword'
        type: array
      popularity:
        type: integer
      producer:
        type: string
      release_year:
        type: integer
      screenshots:
        items:
          $ref: '#/definitions/models.Screenshot'
        type: array
      seasons:
        items:
          $ref: '#/definitions/models.Season'
        type: array
      title:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        maxLength: 50
        minLength: 2
        type: string
      number:
        type: string
      password:
        minLength: 6
        type: string
      refresh_token:
        type: string
      token:
        type: string
      user_id:
        type: string
      user_type:
        type: string
    required:
    - date_of_birth
    - email
    - name
    - number
    - password
    - user_type
    type: object
host: localhost:8080
info:
  contact: {}
  description: Type "JWT" token received from the server
  title: Ozinshe API
  version: "1.0"
paths:
...
  /change-password:
    post:
      consumes:
      - application/json
      description: Allows the currently authenticated user to change their password.
      parameters:
      - description: New password information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.NewPassword'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
        "400":
          description: Error changing password
          schema: {}
      security:
      - CookieAuth: []
      summary: Change password
      tags:
      - users
  /change-profile:
    post:
      consumes:
      - application/json
      description: Allows the currently authenticated user to change their profile
        information.
      parameters:
      - description: New profile information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ProfileData'
      produces:
      - application/json
      responses:
        "200":
          description: Profile changed successfully
        "400":
          description: Error changing profile
          schema: {}
      security:
      - CookieAuth: []
      summary: Change profile
      tags:
      - users
...
  /series:
    get:
      description: Retrieves a list of all series.
      produces:
      - application/json
      responses:
        "200":
          description: List of series
          schema:
            items:
              $ref: '#/definitions/models.Series'
            type: array
        "400":
          description: Error getting series
          schema: {}
      summary: Get a list of all series
      tags:
      - series
  /series/{id}:
    delete:
      description: Deletes a series based on the provided ID.
      parameters:
      - description: Series ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Series deleted successfully
        "400":
          description: Invalid parameters
          schema: {}
        "500":
          description: Error deleting series
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete a series
      tags:
      - series
  /series/{seriesID}:
    get:
      description: Retrieves details of a series based on the provided ID.
      parameters:
      - description: Series ID
        in: path
        name: seriesID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Series details
          schema:
            $ref: '#/definitions/models.Series'
        "400":
          description: Error getting series
          schema: {}
      summary: Get details of a specific series
      tags:
      - series
  /series/{seriesID}/seasons/{seasonNumber}/episodes:
    get:
      description: Retrieves episodes of a specific season of a series based on the
        provided series ID and season number.
      parameters:
      - description: Series ID
        in: path
        name: seriesID
        required: true
        type: string
      - description: Season Number
        in: path
        name: seasonNumber
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of episodes
          schema:
            items:
              $ref: '#/definitions/models.Episode'
            type: array
        "400":
          description: Invalid parameters
          schema: {}
        "404":
          description: Season not found
          schema: {}
        "500":
          description: Error getting episodes
          schema: {}
      summary: Get episodes of a specific season of a series
      tags:
      - series
  /series/{seriesID}/seasons/{seasonNumber}/episodes/{episodeID}:
    get:
      description: Retrieves details of a specific episode of a series based on the
        provided series ID, season number, and episode ID.
      parameters:
      - description: Series ID
        in: path
        name: seriesID
        required: true
        type: string
      - description: Season Number
        in: path
        name: seasonNumber
        required: true
        type: string
      - description: Episode ID
        in: path
        name: episodeID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Episode details
          schema:
            $ref: '#/definitions/models.Episode'
        "400":
          description: Invalid parameters
          schema: {}
        "500":
          description: Error getting episode
          schema: {}
      summary: Get details of a specific episode of a series
      tags:
      - series
...
  /user/{id}:
    delete:
      description: Deletes a user based on the provided ID. Requires admin authorization.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
        "400":
          description: Error deleting user
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete user
      tags:
      - users
    get:
      description: Retrieves details of a user based on the provided ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Error getting user
          schema: {}
      security:
      - CookieAuth: []
      summary: Get details of a specific user
      tags:
      - users
  /users:
    get:
      description: Retrieves a list of all users in the system. Requires admin authorization.
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Error getting users
          schema: {}
      security:
      - CookieAuth: []
      summary: Get a list of all users
      tags:
      - users
swagger: "2.0"